+++ src/ext-test/java/org/opentripplanner/ext/flex/ScheduledDeviatedTripTest.java
@@ -2,19 +2,29 @@ package org.opentripplanner.ext.flex;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.net.URISyntaxException;
+import java.time.OffsetDateTime;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.locationtech.jts.geom.Coordinate;
 import org.opentripplanner.ext.flex.trip.FlexTrip;
 import org.opentripplanner.ext.flex.trip.ScheduledDeviatedTrip;
+import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.FlexStopLocation;
+import org.opentripplanner.routing.api.request.RoutingRequest;
+import org.opentripplanner.routing.core.Fare.FareType;
+import org.opentripplanner.routing.core.Money;
+import org.opentripplanner.routing.core.State;
+import org.opentripplanner.routing.core.WrappedCurrency;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graphfinder.NearbyStop;
+import org.opentripplanner.routing.location.StreetLocation;
+import org.opentripplanner.util.OTPFeature;
 
 /**
  * This tests that the feed for the Cobb County Flex service is processed correctly. This service
@@ -54,7 +64,7 @@ public class ScheduledDeviatedTripTest extends FlexTest {
                 params
         ).collect(Collectors.toList());
 
-        assertEquals(2, accesses.size());
+        assertEquals(3, accesses.size());
 
         var access = accesses.get(0);
         assertEquals(1, access.fromStopIndex);
@@ -73,11 +83,40 @@ public class ScheduledDeviatedTripTest extends FlexTest {
                 params
         ).collect(Collectors.toList());
 
-        assertEquals(2, egresses.size());
+        assertEquals(3, egresses.size());
 
         var egress = egresses.get(0);
-        assertEquals(1, egress.fromStopIndex);
-        assertEquals(1, egress.toStopIndex);
+        assertEquals(2, egress.fromStopIndex);
+        assertEquals(2, egress.toStopIndex);
+    }
+
+    @Test
+    public void calculateDirectFare() {
+        OTPFeature.enableFeatures(Map.of(OTPFeature.FlexRouting, true));
+        var trip = getFlexTrip();
+
+        var from = getNearbyStop(trip, "from-stop");
+        var to = getNearbyStop(trip, "to-stop");
+
+        var router = new FlexRouter(
+                graph,
+                new FlexParameters(300),
+                OffsetDateTime.parse("2021-11-12T10:15:24-05:00").toInstant(),
+                false,
+                1,
+                1,
+                List.of(from),
+                List.of(to)
+        );
+
+        var itineraries = router.createFlexOnlyItineraries();
+
+        var itinerary = itineraries.iterator().next();
+        assertFalse(itinerary.fare.fare.isEmpty());
+
+        assertEquals(new Money(new WrappedCurrency("USD"), 250), itinerary.fare.getFare(FareType.regular));
+
+        OTPFeature.enableFeatures(Map.of(OTPFeature.FlexRouting, false));
     }
 
     @BeforeAll
@@ -86,13 +125,31 @@ public class ScheduledDeviatedTripTest extends FlexTest {
     }
 
     private static NearbyStop getNearbyStop(FlexTrip trip) {
-        var stopLocation = trip.getStops().stream().collect(Collectors.toList()).get(2);
-        assertTrue(stopLocation instanceof FlexStopLocation);
-        return new NearbyStop(stopLocation, 0, List.of(), null, null);
+        return getNearbyStop(trip, "nearby-stop");
+    }
+
+    private static NearbyStop getNearbyStop(FlexTrip trip, String id) {
+        // getStops() returns a set of stops and the order doesn't correspond to the stop times
+        // of the trip
+        var stopLocation = trip.getStops()
+                .stream()
+                .filter(s -> s instanceof FlexStopLocation)
+                .findFirst()
+                .get();
+        var r = new RoutingRequest();
+        r.setRoutingContext(graph);
+        return new NearbyStop(
+                stopLocation,
+                0,
+                List.of(),
+                null,
+                new State(new StreetLocation(id, new Coordinate(0, 0), id), r)
+        );
     }
 
     private static FlexTrip getFlexTrip() {
-        var flexTrips = graph.flexTripsById.values();
-        return flexTrips.iterator().next();
+        var feedId = graph.getFeedIds().iterator().next();
+        var tripId = new FeedScopedId(feedId, "a326c618-d42c-4bd1-9624-c314fbf8ecd8");
+        return graph.flexTripsById.get(tripId);
     }
 }