+++ src/main/java/org/opentripplanner/api/mapping/TripPlanMapper.java
@@ -9,8 +9,8 @@ public class TripPlanMapper {
     private final ItineraryMapper itineraryMapper;
     private final PlaceMapper placeMapper;
 
-    public TripPlanMapper(Locale locale) {
-        this.itineraryMapper = new ItineraryMapper(locale);
+    public TripPlanMapper(Locale locale, boolean addIntermediateStops) {
+        this.itineraryMapper = new ItineraryMapper(locale, addIntermediateStops);
         this.placeMapper = new PlaceMapper(locale);
     }