+++ src/main/java/org/opentripplanner/routing/fares/impl/TimeBasedVehicleRentalFareServiceFactory.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.vehicle_rental;
+package org.opentripplanner.routing.fares.impl;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -10,8 +10,8 @@ import java.util.Map;
 
 import org.opentripplanner.model.OtpTransitService;
 import org.opentripplanner.common.model.P2;
-import org.opentripplanner.routing.services.FareService;
-import org.opentripplanner.routing.services.FareServiceFactory;
+import org.opentripplanner.routing.fares.FareService;
+import org.opentripplanner.routing.fares.FareServiceFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;