+++ src/main/java/org/opentripplanner/updater/alerts/GtfsRealtimeAlertsUpdater.java
@@ -1,20 +1,19 @@
 package org.opentripplanner.updater.alerts;
 
 import com.google.transit.realtime.GtfsRealtime.FeedMessage;
+import java.io.InputStream;
+import java.net.URI;
+import java.util.Map;
 import org.opentripplanner.routing.RoutingService;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.impl.TransitAlertServiceImpl;
 import org.opentripplanner.routing.services.TransitAlertService;
-import org.opentripplanner.updater.GraphUpdaterManager;
 import org.opentripplanner.updater.GtfsRealtimeFuzzyTripMatcher;
 import org.opentripplanner.updater.PollingGraphUpdater;
+import org.opentripplanner.updater.WriteToGraphCallback;
 import org.opentripplanner.util.HttpUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.InputStream;
-import java.net.URI;
-import java.util.Map;
 
 /**
  * GTFS-RT alerts updater
@@ -32,7 +31,7 @@ import java.util.Map;
 public class GtfsRealtimeAlertsUpdater extends PollingGraphUpdater {
     private static final Logger LOG = LoggerFactory.getLogger(GtfsRealtimeAlertsUpdater.class);
 
-    private GraphUpdaterManager updaterManager;
+    private WriteToGraphCallback saveResultOnGraph;
 
     private Long lastTimestamp = Long.MIN_VALUE;
 
@@ -51,8 +50,8 @@ public class GtfsRealtimeAlertsUpdater extends PollingGraphUpdater {
     private TransitAlertService transitAlertService;
 
     @Override
-    public void setGraphUpdaterManager(GraphUpdaterManager updaterManager) {
-        this.updaterManager = updaterManager;
+    public void setGraphUpdaterManager(WriteToGraphCallback saveResultOnGraph) {
+        this.saveResultOnGraph = saveResultOnGraph;
     }
 
     public GtfsRealtimeAlertsUpdater(GtfsRealtimeAlertsUpdaterParameters config) {
@@ -101,7 +100,7 @@ public class GtfsRealtimeAlertsUpdater extends PollingGraphUpdater {
             }
 
             // Handle update in graph writer runnable
-            updaterManager.execute(graph -> updateHandler.update(feed));
+            saveResultOnGraph.execute(graph -> updateHandler.update(feed));
 
             lastTimestamp = feedTimestamp;
         } catch (Exception e) {