+++ src/test/java/org/opentripplanner/routing/street/SplitEdgeTurnRestrictionsTest.java
@@ -18,7 +18,7 @@ import org.opentripplanner.standalone.server.Router;
 import org.opentripplanner.util.TestUtils;
 
 /*
- * When bus stops are added to graph they split an existing edge in two parts so that an artifical
+ * When bus stops are added to graph they split an existing edge in two parts so that an artificial
  * intersection can be added and routes can be found to the station.
  *
  * During this process turn restrictions also need to be propagated from the old edges to the newly
@@ -42,7 +42,7 @@ public class SplitEdgeTurnRestrictionsTest {
 
     @Test
     public void shouldTakeDeufringenTurnRestrictionsIntoAccount() throws IOException {
-        Graph graph = ConstantsForTests.buildGtfsGraph(
+        Graph graph = ConstantsForTests.buildOsmAndGtfsGraph(
                 ConstantsForTests.DEUFRINGEN_OSM,
                 ConstantsForTests.VVS_BUS_764_ONLY
         );
@@ -82,7 +82,7 @@ public class SplitEdgeTurnRestrictionsTest {
     public void shouldTakeBoeblingenTurnRestrictionsIntoAccount() throws IOException {
         // this tests that the following turn restriction is transferred correctly to the split edges
         // https://www.openstreetmap.org/relation/299171
-        var graph = ConstantsForTests.buildGtfsGraph(
+        var graph = ConstantsForTests.buildOsmAndGtfsGraph(
                 ConstantsForTests.BOEBLINGEN_OSM,
                 ConstantsForTests.VVS_BUS_751_ONLY
         );